Adjust Day 25 output cast

+++ 2024/25/a.cpp
@@ -35,5 +35,5 @@ int main() {
     return (lock & key) == 0;
   };
 
-  cout << ranges::count_if(views::cartesian_product(locks, keys), isValidPair) << endl;
+  cout << static_cast<uint64_t>(ranges::count_if(views::cartesian_product(locks, keys), isValidPair)) << endl;
 }
